move user and post api calls under api-service folder
