+++ src/hooks/useUser.ts
@@ -8,12 +8,12 @@ import { onAuthStateChanged } from 'firebase/auth';
 import { auth } from '@/services/firebase-service/firebase-config';
 
 import { queryClient } from '@/components/tanstack-provider/tanstack-provider';
-import userService from '@/services/user-service/user-service';
 import { setIsAuthLoading, setIsAuthModalOpen } from '@/store/reducers/ui-reducer/ui-slice';
 import { UserState, setUser } from '@/store/reducers/user-reducer/user-slice';
 import userReducerSelector from '@/store/reducers/user-reducer/user-reducer-selector';
 
 import { User } from '@/services/firebase-service/types/db-types/user';
+import userApiService from '@/services/api-service/user-api-service/user-api-service';
 
 const useUser = (): UserState => {
   const dispatch = useDispatch();
@@ -23,7 +23,7 @@ const useUser = (): UserState => {
   useEffect(() => {
     const unsubscribe = onAuthStateChanged(auth, async user => {
       if (user) {
-        const userData = (await userService.getUserById(user.uid)) as User;
+        const userData = (await userApiService.getUserById(user.uid)) as User;
 
         // register case
         if (!userData) {
@@ -59,7 +59,7 @@ const useUser = (): UserState => {
         );
         await queryClient.fetchQuery({
           queryKey: ['logOutUser'],
-          queryFn: () => userService.logOutUser(),
+          queryFn: () => userApiService.logOutUser(),
         });
         router.push('/');
       }
